+++ modules/@angular/platform-browser/animations/src/animations.ts
@@ -11,7 +11,6 @@
  * @description
  * Entry point for all animation APIs of the animation browser package.
  */
-export {BrowserAnimationsModule} from './browser_animations_module';
-export {NoopAnimationsModule} from './noop_animations_module';
+export {BrowserAnimationsModule, NoopAnimationsModule} from './module';
 export {AnimationDriver} from './render/animation_driver';
 export * from './private_export';
