+++ modules/@angular/platform-browser/animations/src/providers.ts
@@ -5,8 +5,9 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {Injectable, NgModule, NgZone, RendererFactoryV2} from '@angular/core';
-import {BrowserModule, ɵDomRendererFactoryV2} from '@angular/platform-browser';
+
+import {Injectable, NgZone, Provider, RendererFactoryV2} from '@angular/core';
+import {ɵDomRendererFactoryV2} from '@angular/platform-browser';
 
 import {AnimationEngine} from './animation_engine';
 import {AnimationStyleNormalizer} from './dsl/style_normalization/animation_style_normalizer';
@@ -14,6 +15,7 @@ import {WebAnimationsStyleNormalizer} from './dsl/style_normalization/web_animat
 import {AnimationDriver, NoopAnimationDriver} from './render/animation_driver';
 import {AnimationRendererFactory} from './render/animation_renderer';
 import {DomAnimationEngine} from './render/dom_animation_engine';
+import {NoopAnimationEngine} from './render/noop_animation_engine';
 import {WebAnimationsDriver, supportsWebAnimations} from './render/web_animations/web_animations_driver';
 
 @Injectable()
@@ -40,19 +42,27 @@ export function instantiateRendererFactory(
 }
 
 /**
- * @experimental Animation support is experimental.
+ * Separate providers from the actual module so that we can do a local modification in Google3 to
+ * include them in the BrowserModule.
  */
-@NgModule({
-  imports: [BrowserModule],
-  providers: [
-    {provide: AnimationDriver, useFactory: instantiateSupportedAnimationDriver},
-    {provide: AnimationStyleNormalizer, useFactory: instantiateDefaultStyleNormalizer},
-    {provide: AnimationEngine, useClass: InjectableAnimationEngine}, {
-      provide: RendererFactoryV2,
-      useFactory: instantiateRendererFactory,
-      deps: [ɵDomRendererFactoryV2, AnimationEngine, NgZone]
-    }
-  ]
-})
-export class BrowserAnimationsModule {
-}
+export const BROWSER_ANIMATIONS_PROVIDERS: Provider[] = [
+  {provide: AnimationDriver, useFactory: instantiateSupportedAnimationDriver},
+  {provide: AnimationStyleNormalizer, useFactory: instantiateDefaultStyleNormalizer},
+  {provide: AnimationEngine, useClass: InjectableAnimationEngine}, {
+    provide: RendererFactoryV2,
+    useFactory: instantiateRendererFactory,
+    deps: [ɵDomRendererFactoryV2, AnimationEngine, NgZone]
+  }
+];
+
+/**
+ * Separate providers from the actual module so that we can do a local modification in Google3 to
+ * include them in the BrowserTestingModule.
+ */
+export const BROWSER_NOOP_ANIMATIONS_PROVIDERS: Provider[] = [
+  {provide: AnimationEngine, useClass: NoopAnimationEngine}, {
+    provide: RendererFactoryV2,
+    useFactory: instantiateRendererFactory,
+    deps: [ɵDomRendererFactoryV2, AnimationEngine, NgZone]
+  }
+];
